requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==23.2.0
Babel==2.10.3
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.7.0
Bottleneck==1.3.5
Brotli==1.1.0
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
click==8.1.6
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
contourpy==1.3.2
cryptography==41.0.7
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
et-xmlfile==1.0.1
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.7.0
fs==2.4.16
git-filter-repo==2.38.0
greenlet==3.2.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httplib2==0.20.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.6
iniconfig==1.1.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.5
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==0.0.0
multidict==6.6.3
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook_shim==0.2.4
numexpr==2.9.0
numpy==2.2.5
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openpyxl==3.1.2
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.1.4+dfsg
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.7
pluggy==1.4.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
pytest==7.4.4
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==26.4.0
referencing==0.36.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.24.0
scikit-learn==1.4.1.post1
scipy==1.15.3
Send2Trash==1.8.3
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.12
systemd-python==235
tables==3.9.2
tenacity==9.1.2
terminado==0.18.1
terminator==2.1.3
threadpoolctl==3.1.0
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ufoLib2==0.16.0
unicodedata2==15.1.0
uri-template==1.3.0
urllib3==2.0.7
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.42.0
xdg==5
yarl==1.20.1
zstandard==0.23.0
python-dotenv>=1.0.0
langchain-qdrant
pypdf>=3.0
qdrant-client>=1.7.0
langchain>=0.1.17
langchain_qdrant>=0.1.3
langchain_ollama>=0.1.0
